Flytt funksjoner til common-mappe
