+++ scripts/dask_infrastructure.py
@@ -1,21 +1,6 @@
 import sys
 import json
-import re
-
-def replace_special_characters(s):
-    replacements = {
-        'æ': 'ae',
-        'Æ': 'Ae',
-        'ø': 'o',
-        'Ø': 'O',
-        'å': 'aa',
-        'Å': 'Aa'
-    }
-    
-    def replace(match):
-        return replacements[match.group(0)]
-    
-    return re.sub(r'æ|Æ|ø|Ø|å|Å', replace, s)
+from common import replace_special_characters, append_content_to_end_of_file
 
 def edit_file(filepath, params):
     project_name: str = params.get("project_name")
@@ -70,17 +55,6 @@ def generate_module_definition(ad_group_name: str, area_name: str, project_name:
     '''
 
     return module
-
-def append_content_to_end_of_file(file_path: str, content: str) -> None:
-    with open(file_path) as file:
-        lines = file.readlines()
-        file.close()
-        lines.insert(len(lines), content)
-
-        with open(file_path, 'w') as file:
-            file.writelines(lines)
-            file.close()
-
 
 
 if __name__ == "__main__":
